+++ source/renderer/app/api/nodes/requests/postponeAppUpdate.js
@@ -2,9 +2,9 @@
 import type { RequestConfig } from '../../common/types';
 import { request } from '../../utils/request';
 
-export const getNextNodeUpdate = (config: RequestConfig): Promise<any> =>
+export const postponeAppUpdate = (config: RequestConfig): Promise<any> =>
   request({
-    method: 'GET',
-    path: '/api/internal/next-update',
+    method: 'POST',
+    path: '/api/internal/postpone-update',
     ...config,
   });