+++ source/renderer/app/containers/loading/SyncingConnectingPage.js
@@ -34,7 +34,7 @@ export default class LoadingSyncingConnectingPage extends Component<Props> {
       isNewAppVersionAvailable,
       isNewAppVersionLoading,
       isNewAppVersionLoaded,
-    } = stores.nodeUpdate;
+    } = stores.appUpdate;
     const { hasLoadedCurrentLocale, hasLoadedCurrentTheme } = stores.profile;
     const { toggleNewsFeed } = this.props.actions.app;
     const { unread } = stores.newsFeed.newsFeedData;
@@ -101,8 +101,8 @@ export default class LoadingSyncingConnectingPage extends Component<Props> {
   };
 
   handleGetAvailableVersions = () => {
-    const { nodeUpdate } = this.props.actions;
-    nodeUpdate.getLatestAvailableAppVersion.trigger();
+    const { appUpdate } = this.props.actions;
+    appUpdate.getLatestAvailableAppVersion.trigger();
   };
 
   openDaedalusDiagnosticsDialog = () => {