+++ source/renderer/app/stores/index.js
@@ -3,9 +3,9 @@ import { observable, action } from 'mobx';
 import type Store from './lib/Store';
 import AddressesStore from './AddressesStore';
 import AppStore from './AppStore';
+import AppUpdateStore from './AppUpdateStore';
 import NetworkStatusStore from './NetworkStatusStore';
 import NewsFeedStore from './NewsFeedStore';
-import NodeUpdateStore from './NodeUpdateStore';
 import ProfileStore from './ProfileStore';
 import SidebarStore from './SidebarStore';
 import StakingStore from './StakingStore';
@@ -22,9 +22,9 @@ import WindowStore from './WindowStore';
 export const storeClasses = {
   addresses: AddressesStore,
   app: AppStore,
+  appUpdate: AppUpdateStore,
   networkStatus: NetworkStatusStore,
   newsFeed: NewsFeedStore,
-  nodeUpdate: NodeUpdateStore,
   profile: ProfileStore,
   sidebar: SidebarStore,
   staking: StakingStore,
@@ -42,9 +42,9 @@ export const storeClasses = {
 export type StoresMap = {
   addresses: AddressesStore,
   app: AppStore,
+  appUpdate: AppUpdateStore,
   networkStatus: NetworkStatusStore,
   newsFeed: NewsFeedStore,
-  nodeUpdate: NodeUpdateStore,
   profile: ProfileStore,
   router: Object,
   sidebar: SidebarStore,
@@ -86,7 +86,7 @@ export default action((api, actions, router): StoresMap => {
     app: createStoreInstanceOf(AppStore),
     networkStatus: createStoreInstanceOf(NetworkStatusStore),
     newsFeed: createStoreInstanceOf(NewsFeedStore),
-    nodeUpdate: createStoreInstanceOf(NodeUpdateStore),
+    appUpdate: createStoreInstanceOf(AppUpdateStore),
     profile: createStoreInstanceOf(ProfileStore),
     router,
     sidebar: createStoreInstanceOf(SidebarStore),