+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog
 
 ### Features
 
+- Implemented a download manager ([PR 2020](https://github.com/input-output-hk/daedalus/pull/2020))
 - Implemented Hardware wallets connection screens ([PR 2016](https://github.com/input-output-hk/daedalus/pull/2016))
 
 ### Fixes